Fix build and lint warnings
stackoverflow:
https://stackoverflow.com/questions/44322178/room-schema-export-directory-is-not-provided-to-the-annotation-processor-so-we

"Schema export directory is not provided to the annotation processor so
we cannot export the schema. You can either provide  room.schemaLocation
annotation processor argument OR set exportSchema to false."

You can set annotation processor argument (room.schemaLocation) to tell
Room to export the schema into a folder. Even though it is not
mandatory, it is a good practice to have version history in your
codebase and you should commit that file into your version control
system (but don't ship it with your app!).

So if you don't need to check the schema and you want to get rid of the
warning, just add exportSchema = false to your RoomDatabase, as follows.

@Database(entities = { YourEntity.class }, version = 1, exportSchema =
false)
public abstract class AppDatabase extends RoomDatabase {
   //...
}